requirements.txt
requests
beautifulsoup4
pandas
numpy
psycopg2-binary
python-dotenv